Add Top Bar and Filters Bar
Adds Top Bar and Filters Bar to the Search Page. The Top Bar contains a title and a button that redirects to the suggest a new service page. The Filters Bar contains a search bar and four filters for: service type, taxonomic scope, publishing date and language. The filters will need to be made functional in the future, but the state foundation with Formik is there.