Search Result(s)
Adds Components and fetch functionality for displaying Search Results. The Search Results component calls on the taxonomic services from the API (or mock) and then renders a Search Result component for each result. The Search Result component's lay-out changes depending on if there is a preview image present or not.

This commit also adds a couple of mock data files for testing as long as the API is not up.