Add Hooks for Pagination and Fetch to handle these outside of normal files
Adds Custom Hooks for Pagination and Fetch functions. These will handle the async calls behind the scenes and can be reused. It cleans up the code on the main files like the Search page. It is still necessary, for the global redux state, to pass on a function to update this. Can be smoothed in the future, but already makes life a bit easier.

Also adds Spinner when content is loading.