Add Reference collections look, make filters function and add expertise/404 page
Adds the reference collections look when selecting reference collections on the home page or using the search filter. It will make the page background and elements green, according to the color of the reference collections.

To suffice this the search filters have been made functional, adding their values to the search params.

Also added a blank page for the taxonomic expertise and a 404 page.