Integrate API responses into the Marketplace
Adds support for API responses.
The Marketplace now functions and relies upon responses from the Cordra API. Search results, individual detail pages and numbers on the homepage will only work if the API is online. If the API is offline, a message will be shown in the header and no results will be displayed. The Marketplace is still interactive and prevents users from executing breaking actions while in this state.