Small improvements based upon API
Adds small improvements based on the API.

Makes sure all of the filters can be reset and revert to their original values and inherit the search params. Error handling for empty responses with a 200 for search and the taxonomic page have been improved to know show custom messages and adapted behavior.